Oops all thieves (#4786)
* new gamemode: OopsAllThieves

* change amount of thieves rolled as per maintainer discussion, add to secret_weights.yml

* add traitors subgamemode

* add traitors subgamemode

* Cleanup and light tuning once again.

This time we increase traitors limit to 3, and make it a guarantee that traitors is included in an OopsAllThieves gamemode, as we really need the chaos (seen in in-game testing).

Absolute, worst case, lowpop scenario, out of 30 players, 3 will be a traitor, and 4 will be a thief, meaning that there will be 23 non-antags, or ~76% of players will be nonantag (~23% will be antag).

Considering this is all at a 5% chance I think it's fine. This can see tuning on Vulture.

* didn't see that

* Added to secret weights

* goobcomment

* Removed dupicate

* Fixes

* Fixed linux

* Goobstation namespace

---------

Co-authored-by: ArtisticRoomba <145879011+ArtisticRoomba@users.noreply.github.com>

(cherry picked from commit 45eb7b4ff6a62684c01221d5416a9407b4ed6cc1)